First prototype of creating and changing scenes dynamically
The menu scene creates four scenes in a 2 x 2 grid with names 'level_x_y' and then immediately starts 'level_0_0'.
Use arrow keys to move to different scenes.
Each scene is given a random background color so you can tell when the scene has changed.
The game object now has 2 new fields:
mapDimensions - a Vector2 that holds the dimensions of the whole map
map - a 2d array with the levels held in it (this actually may not be necessary)